@@ -1,28 +1,31 @@
 {
-  "name": "react-auth-template",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.4",
+    "@fortawesome/fontawesome-free": "^6.5.1",
     "@headlessui/react": "^1.7.17",
     "@heroicons/react": "^2.1.1",
-    "@popperjs/core": "^2.10.2",
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "bootstrap": "^5.1.3",
+    "@popperjs/core": "^2.11.8",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "axios": "^1.6.2",
+    "bootstrap": "^5.3.2",
     "flowbite": "^2.2.1",
     "flowbite-react": "^0.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^5.0.1",
-    "sass": "^1.43.4",
-    "web-vitals": "^1.1.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.21.1",
+    "react-scripts": "^3.0.1",
+    "sass": "^1.69.5",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build"
+    "start": "react-scripts --openssl-legacy-provider start ",
+    "build": "react-scripts --openssl-legacy-provider start",
+    "test": "react-scripts --openssl-legacy-provider start",
+    "eject": "react-scripts --openssl-legacy-provider start"
   },
   "eslintConfig": {
     "extends": [
@@ -44,6 +47,5 @@
   },
   "devDependencies": {
     "tailwindcss": "^3.4.0"
-  },
-  "proxy": "http://127.0.0.1:8000"
+  }
 }
