@@ -1,16 +1,17 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+import './assets/scss/auth.scss'
+import {CartProvider} from './utilities/cart'
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import './assets/scss/auth.scss'
-import Header from "./components/views/Header";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Header />
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+    <React.StrictMode>
+        <CartProvider>
+            <App/>
+        </CartProvider>
+    </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
