@@ -1,8 +1,7 @@
 from django.urls import path, include
 from rest_framework import routers
 
-from clothing_store.views import HatViewSet, FurCoatViewSet, GlovesViewSet, BagViewSet, PhotoFurCoatView, PhotoBagView, \
-    PhotoGlovesView, PhotoHatView
+from clothing_store.views import HatViewSet, FurCoatViewSet, GlovesViewSet, BagViewSet
 
 router = routers.SimpleRouter()
 
@@ -13,8 +12,4 @@ router.register(r'bag', BagViewSet)
 
 urlpatterns = [
     path('', include(router.urls)),
-    path('photos-furcoat', PhotoFurCoatView.as_view(), name='photos_furcoat'),
-    path('photos-bag', PhotoBagView.as_view(), name='photos_bag'),
-    path('photos-gloves', PhotoGlovesView.as_view(), name='photos_gloves'),
-    path('photos-hat', PhotoHatView.as_view(), name='photos_hat'),
 ]
