@@ -1,3 +1,6 @@
+import inspect
+import pprint
+import sys
 from uuid import UUID
 
 from django.http import JsonResponse
@@ -7,8 +10,10 @@ from rest_framework.views import APIView
 from django.db.models import Model
 from rest_framework.request import Request
 
+from backend.settings import Products
+from clothing_store import models, serializers
 from clothing_store.models import FurCoat, Hat, Bag, Gloves, AbstractProduct, PhotoFurCoat, PhotoGloves, PhotoHat, \
-    PhotoBag
+    PhotoBag, Product
 from clothing_store.serializers import FurCoatSerializer, GlovesSerializer, HatSerializer, BagSerializer, \
     PhotoFurCoatSerializer, PhotoGlovesSerializer, PhotoHatSerializer, PhotoBagSerializer
 
@@ -17,53 +22,69 @@ class FurCoatViewSet(viewsets.ModelViewSet):
     queryset = FurCoat.objects.all()
     serializer_class = FurCoatSerializer
 
+    def list(self, request, *args, **kwargs):
+        fur_coats = FurCoat.objects.all()
+        data = []
+        for fur_coat in fur_coats:
+            photos = list()
+            photos_object = PhotoFurCoat.objects.filter(fur_coat=fur_coat)
+            photos.append(PhotoFurCoatSerializer(photos_object, many=True).data)
+            data.append({
+                **FurCoatSerializer(fur_coat).data,
+                "photos": photos
+            })
+        return JsonResponse(data, safe=False)
+
 
 class HatViewSet(viewsets.ModelViewSet):
     queryset = Hat.objects.all()
     serializer_class = HatSerializer
+
+    def list(self, request, *args, **kwargs):
+        hats = Hat.objects.all()
+        data = []
+        for hat in hats:
+            photos = list()
+            photos_object = PhotoHat.objects.filter(hat=hat)
+            photos.append(PhotoHatSerializer(photos_object, many=True).data)
+            data.append({
+                **HatSerializer(hat).data,
+                "photos": photos
+            })
+        return JsonResponse(data, safe=False)
 
 
 class BagViewSet(viewsets.ModelViewSet):
     queryset = Bag.objects.all()
     serializer_class = BagSerializer
 
+    def list(self, request, *args, **kwargs):
+        bags = Bag.objects.all()
+        data = []
+        for bag in bags:
+            photos = list()
+            photos_object = PhotoBag.objects.filter(bag=bag)
+            photos.append(PhotoBagSerializer(photos_object, many=True).data)
+            data.append({
+                **BagSerializer(bag).data,
+                "photos": photos
+            })
+        return JsonResponse(data, safe=False)
+
 
 class GlovesViewSet(viewsets.ModelViewSet):
     queryset = Gloves.objects.all()
     serializer_class = GlovesSerializer
 
-
-class PhotoFurCoatView(APIView):
-    def get(self, request: Request, *args, **kwargs):
-        data = request.data
-        fur_coat = get_object_or_404(FurCoat, pk=data.get('id'))
-        photos = PhotoFurCoat.objects.filter(fur_coat=fur_coat)
-        photos_serializer = PhotoFurCoatSerializer(photos, many=True)
-        return JsonResponse(photos_serializer.data, safe=False)
-
-
-class PhotoGlovesView(APIView):
-    def get(self, request: Request, *args, **kwargs):
-        data = request.data
-        gloves = get_object_or_404(Gloves, pk=data.get('id'))
-        photos = PhotoGloves.objects.filter(gloves=gloves)
-        photos_serializer = PhotoGlovesSerializer(photos, many=True)
-        return JsonResponse(photos_serializer.data, safe=False)
-
-
-class PhotoHatView(APIView):
-    def get(self, request: Request, *args, **kwargs):
-        data = request.data
-        hat = get_object_or_404(Hat, pk=data.get('id'))
-        photos = PhotoHat.objects.filter(hat=hat)
-        photos_serializer = PhotoHatSerializer(photos, many=True)
-        return JsonResponse(photos_serializer.data, safe=False)
-
-
-class PhotoBagView(APIView):
-    def get(self, request: Request, *args, **kwargs):
-        data = request.data
-        bag = get_object_or_404(Bag, pk=data.get('id'))
-        photos = PhotoBag.objects.filter(bag=bag)
-        photos_serializer = PhotoBagSerializer(photos, many=True)
-        return JsonResponse(photos_serializer.data, safe=False)
+    def list(self, request, *args, **kwargs):
+        gloves = Gloves.objects.all()
+        data = []
+        for glove in gloves:
+            photos = list()
+            photos_object = PhotoGloves.objects.filter(gloves=glove)
+            photos.append(PhotoGlovesSerializer(photos_object, many=True).data)
+            data.append({
+                **GlovesSerializer(glove).data,
+                "photos": photos
+            })
+        return JsonResponse(data, safe=False)
